BGDIINF_SB-1799 raise error in unit tests
in case the bucket and the table that are mocked in the unit test already exist, an error is raised.
As the bucket and the table are deleted in the tearDown(), there should not be any existing bucket
and table in setUp(). This also makes sure that we do not accidentally use real resources.