fixed broken unit test
In PR https://github.com/geoadmin/service-kml/pull/14 a cline was deleted from a unit test by mistake.
This line was part of the test_valid_kml_put() and casserted the equality of response_post.json and response.json.
However, since the updated timestamp was added to the responses, this assertion would have failed. So the
assertion was re-added in this PR but ignoring the updated key of the responses' jsons, as the updated timestamps
should and may differ.