Removed the route prefix and use SCRIPT_NAME, configure gunicorn for secure scheme header
The path prefix is now set by the deployment using the WSGI SCRIPT_NAME
environment variable.

Also configure gunicorn to use secure scheme header which allow us to
get rid of the middleware. This settings is needed in order to set the
correct http proto in generated urls in responses.

Change the default ALLOWED_ORIGIN to allow everything.