Fixed CORS allowed method
The allowed method in CORS header where not correct, the list contained
only the methods for the endpoint (methods registered by the python
route function). For example the path GET /admin/<kml_id> returned the
list GET, HEAD, OPTIONS, but the PUT and DELETE where missing because
they are part of other Flask endpoints.

Also added the ALLOWED_DOMAINS in .env.default in order to simplify the
e2e test with localhost.