Normalized routes logging
Most of the services uses app.routes logger for routes logging. So use it here
too. Also the incoming requests are mostly logged as DEBUG while the outgoing
requests (response) are logged as INFO in most of the services. This allows for
example to filter in Kibana to have only successful requests and/or kind of
access logs.