BGDIINF_SB-2346: Only check the kml file size and not the payload size
The payload size can change depending on the method POST or PUT and also
depending on the client that sets the mutlipart boundary string.

Therefore to avoid issue with kml file that might once be created and thereafter
not be edited anymore due to the overhead of the multipart payload we changed
the logic by checking the size of the file instead of the payload.