BGDIINF_SB-2346: Avoid memory starting
The current kml file size limit doesn't prevent a memory starving issue
as this will first load the file into memory before doing the size check.

So we add a payload check with a high enough multipart/form-data overhead before
reading/uploading the data, so the request can be rejected before starving
the memory.

Note it is not safe to use the KML_MAX_SIZE as the payload max size, because
the payload overhead size changes between POST and PUT, so we might create files
but not modified them anymore even if the size did not changed.